fix(gpu): align prover stack with upstream contracts (#371)
## What ❔

- Read unified circuit/setup metadata from canonical upstream
definitions instead of duplicated GPU constants.
- Align GKR/WHIR proof layout and transcript contracts, unified fixture
geometry, proof-slab alignment, and Rust↔CUDA drift guards.

## Why ❔

Upstream changes landed without corresponding GPU updates:

- [#365](https://github.com/matter-labs/zksync-airbender/pull/365)
changed circuit, GKR, proof/transcript, and canonical metadata
contracts.
- [#368](https://github.com/matter-labs/zksync-airbender/pull/368)
regenerated circuit artifacts while GPU code still assumed the previous
values.

This restores CPU/GPU parity and ties future changes to canonical
owners.

Validated with 8/8 scoped checks, 13/13 default GPU tests, and 76/76
complete GPU tests.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted.

+++ gpu/ntt/src/ntt/tests/dit_engine.rs
@@ -175,7 +175,7 @@ fn run_single_pass_stream_parity(log_n: u32, log_vpt: u32) {
     let n: usize = 1 << log_n;
     let num_cosets = single_pass_num_cosets(log_n, log_vpt);
     // log_lde_factor satisfies num_cosets = 2^log_lde_factor.
-    let log_lde_factor = num_cosets.trailing_zeros() as u32;
+    let log_lde_factor = num_cosets.trailing_zeros();
     assert_eq!(1usize << log_lde_factor, num_cosets);
 
     // coset_step = 2^(OMEGA_LOG_ORDER - log_n - log_lde_factor)
